Move and rename all stories to better structure files
- Move files into appropriate subfolders
- Update sidebar to have an more intuitive structure based on ux
