+++ stories/Introduction.mdx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/blocks'
 
-<Meta title="Examples/Introduction" />
+<Meta title="Introduction" />
 
 <style>
     {`
