+++ stories/SpotPrice/Overview.mdx
@@ -1,11 +1,11 @@
 import { Canvas, Meta, Title } from '@storybook/blocks'
 import * as AdviceGraphStories from './AdviceGraph.stories'
 
-<Meta of={AdviceGraphStories} name="Documentation" />
+<Meta name="Spot Price Documentation" />
 
-<Title>Advice Graph Docs</Title>
+<Title>Spot Price Docs</Title>
 
-This component can be used to render the Advice Graph with data from the API.
+This component can be used to render the Spot Price graphs with data from the API.
 
 The default view is as follows:
