+++ stories/forecast/Overview.mdx
@@ -1,0 +1,14 @@
+import { Canvas, Meta, Title } from '@storybook/blocks'
+import * as ForecastTableStories from './ForecastTable.stories'
+
+<Meta name="Forecast Documentation" />
+
+<Title>Forecast Docs</Title>
+
+This component can be used to render the Forecast tables with data from the API.
+
+The default view is as follows:
+
+<Canvas of={ForecastTableStories.Default} />
+
+## Use Case
