+++ stories/settings/Overview.mdx
@@ -1,0 +1,14 @@
+import { Canvas, Meta, Title } from '@storybook/blocks'
+import * as CarSettingsStories from './CarSettings.stories'
+
+<Meta name="Settings Documentation" />
+
+<Title>Settings Docs</Title>
+
+This component can be used to render the Settings with your custom car data.
+
+The default view is as follows:
+
+<Canvas of={CarSettingsStories.Default} />
+
+## Use Case
